api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-cloud-bigquery>=3.20.0
pydantic>=2.0.0
